chore(test): Allow testing on Python 3.5 and 3.7
Python 3.5 is the minimum required version of storyscript,
which depends on story-hub.

Add py35 and py37 to tox envlist, and removes test deps pins
as the appropriate versions for each Python version should
be determined by the pytest packaging as these are very
well maintained plugins.

Also limit coverage reports to the current module,
removing everything in .tox from the reports.

And do not install library and deps into pep8 test environment.